# openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Recovy",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://recovy.herokuapp.com/api/v1"
    }
  ],
  "paths": {
...
    "/discussion": {
      "post": {
        "tags": ["Discussion"],
        "summary": "Membuat dicussion baru",
        "description": "Membuat dicussion baru memerlukan access token dari author",
        "parameters": [
          {
            "in": "header",
            "description": "Wajib mencamtumkan access token pada Header dengan format \"Authorization: bearer [token] \"",
            "name": "Authorization",
            "required": true,
            "schema": {
              "type": "string",
              "example": "bearer your_token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "authorID": {
                    
                    "type": "integer",
                    "minimum": 1
                  },
                  "picture": {
                  
                    "type": "string",
                    "maximum": 512
                  },
                  "title": {
                    
                    "type": "string",
                    "maximum": 255
                  },
                  "description": {
                    "type": "string"
                    
                  },
                  "category": {
                    "type": "string"
                    
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Jika sukses akan selalu mengembalikan object pada properti data",
            "content": {
              "application/json": {
                "schema": {
                 
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "default": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                         
                          "type": "integer",
                          "minimum": 1
                        },
                        "authorID": {
                         
                          "type": "integer",
                          "minimum": 1
                        },
                        "picture": {
                       
                          "type": "string",
                          "maximum": 512
                        },
                        "title": {
                          
                          "type": "string",
                          "maximum": 255
                        },
                        "description": {
                          "type": "string"
                        },
                        "category": {
                          "type": "string",
                        
                          "maximum": 128
                        },
                        "createdAt": {
                          "type": "integer",
                        
                          "minimum": 1
                        },
                        "updatedAt": {
                          "type": "integer",
                         
                          "minimum": 1
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/discussion/{discussionID}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "description": "ID dari article",
            "required": true,
            "name": "webinarID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "tags": ["Discussion"],
        "summary": "Get dicussion berdasar id-nya",
        "responses": {
          "200": {
            "description": "Jika sukses akan selalu mengembalikan object pada properti data",
            "content": {
              "application/json": {
                "schema": {
            
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "default": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                         
                          "type": "integer",
                          "minimum": 1
                        },
                        "author": {
                          
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "number"
                            
                            },
                            "name": {
                              "type": "string"
                            
                            },
                            "photo": {
                              "type": "string"
                            }
                          }
                        },
                        "picture": {
                      
                          "type": "string",
                          "maximum": 512
                        },
                        "title": {
                         
                          "type": "string",
                          "maximum": 255
                        },
                        "description": {
                          "type": "string"
                        },
                        "category": {
                          "type": "string",
                         
                          "maximum": 128
                        },
                        "createdAt": {
                          "type": "integer",
                        
                          "minimum": 1
                        },
                        "updatedAt": {
                          "type": "integer",
                        
                          "minimum": 1
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Discussion"],
        "summary": "Menghapus article",
        "description": "Menghapus article memerlukan access token dari author atau pemilik. Jika menghapus item dari bukan miliknya akan mengembalikan error forbidden",
        "parameters": [
          {
            "in": "header",
            "description": "Wajib mencamtumkan access token pada Header dengan format \"Authorization: bearer [token] \"",
            "name": "Authorization",
            "required": true,
            "schema": {
              "type": "string",
              "example": "bearer your_token"
            }
          },
          {
            "required": true,
            "description": "ID dari article",
            "in": "path",
            "name": "articleID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Jika sukses mengahpus, akan selalu mengembalikan properti success yang bernilai true",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "default": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/discussions/{category}": {
      "get": {
        "tags": ["Discussion"],
        "summary": "Mendapatkan kumpulan discussions berdasarkan category",
        "parameters": [
          {
            "in": "path",
            "name": "category",
            "example": "health",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Jika sukses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "default": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                         
                            "type": "integer",
                            "minimum": 1
                          },
                          "author": {
                     
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "number"
                          
                              },
                              "name": {
                                "type": "string"
                     
                              },
                              "photo": {
                                "type": "string"
                              }
                            }
                          },
                          "picture": {
                      
                            "type": "string",
                            "maximum": 512
                          },
                          "title": {
                        
                            "type": "string",
                            "maximum": 255
                          },
                          "description": {
                            "type": "string"
                          },
                          "category": {
                            "type": "string",
                   
                            "maximum": 128
                          },
                          "createdAt": {
                            "type": "integer",
                      
                            "minimum": 1
                          },
                          "updatedAt": {
                            "type": "integer",
             
                            "minimum": 1
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/discussion-comment": {
      "post": {
        "tags": ["Discussion Comment"],
        "description": "Menambahkan comment pada discussion",
        "parameters": [
          {
            "in": "header",
            "description": "Wajib mencamtumkan access token pada Header dengan format \"Authorization: bearer [token] \"",
            "name": "Authorization",
            "required": true,
            "schema": {
              "type": "string",
              "example": "bearer your_token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
   
                "properties": {
                  "discussionID": {
                    "type": "integer",
                    "minimum": 1
  
                  },
                  "description": {
                    "type": "string"
       
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Jika sukses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
     
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "default": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
              
                          "minimum": 1
                        },
                        "discussionID": {
                          "type": "integer",
                        
                          "minimum": 1
                        },
                        "commentatorID": {
                          "type": "integer",
                     
                          "minimum": 1
                        },
                        "description": {
                          "type": "string"
                     
                        },
                        "createdAt": {
                          "type": "integer",
                      
                          "minimum": 1
                        },
                        "updatedAt": {
                          "type": "integer",
                         
                          "minimum": 1
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/discussion-comment/{discussionCommentID}": {
      "get": {
        "tags": ["Discussion Comment"],
        "description": "Mendapatkan komentar berdasar id",
        "responses": {
          "200": {
            "description": "Jika sukses",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "default": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                        
                          "minimum": 1
                        },
                        "discussionID": {
                          "type": "integer",
                 
                          "minimum": 1
                        },
                        "commentatorID": {
                          "type": "integer",
                          
                          "minimum": 1
                        },
                        "description": {
                          "type": "string"
                        
                        },
                        "createdAt": {
                          "type": "integer",
                          
                          "minimum": 1
                        },
                        "updatedAt": {
                          "type": "integer",
                         
                          "minimum": 1
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Discussion Comment"],
        "description": "Menghapus comment pada discussion",
        "parameters": [
          {
            "in": "header",
            "description": "Wajib mencamtumkan access token pada Header dengan format \"Authorization: bearer [token] \"",
            "name": "Authorization",
            "required": true,
            "schema": {
              "type": "string",
              "example": "bearer your_token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Jika sukses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "default": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/discussion-comments/{discussionID}": {
      "get": {
        "tags": ["Discussion Comment"],
        "description": "Mendapatkan kumpulan comments berdasarkan discussionID",
        "parameters": [
          {
            "required": true,
            "in": "path",
            "name": "discussionID",
            "description": "ID dari discussion. karena discussion dapat memiliki lebih dari 1 komentar maka akan mengembalikan array",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Jika sukses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "default": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                  
                            "minimum": 1
                          },
                          "discussionID": {
                            "type": "integer",
                  
                            "minimum": 1
                          },
                          "commentatorID": {
                            "type": "integer",
                 
                            "minimum": 1
                          },
                          "description": {
                            "type": "string"
               
                          },
                          "createdAt": {
                            "type": "integer",
                  
                            "minimum": 1
                          },
                          "updatedAt": {
                            "type": "integer",
                    
                            "minimum": 1
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}